INTEL_DII: drm/i915/debugger: Split heartbeat and hangcheck by adding attention scan
By disabling the hangcheck heatbeat is stopped. Introduction of the EU
Debugger raised the need to periodically scan for attention bits
regardless of the hangcheck. The simplest way is to leverage heartbeat
for this but split heartbeats and hangcheck.

This patch changes the heartbeat beahavior. Now it ticks if engine is
awake regardless of the hangcheck. If hangcheck is disabled then only
check for attention will be performed.

If EU debugger is connected then heartbeat interval is automatically
changed to 100ms (hardcoded for now) to allow quick reactions while
debugging. This concerns only RCS/CCS engines. Other engines are not
affected.

It is recommended to have hangcheck disabled for EU Debugger.

Note: it is still possible to stop heartbeat by writing 0 to engine's
heartbeat_interval_ms. It will stop heartbeat but only when EU
Debugger is not connected.

Signed-off-by: Maciej Patelczyk <maciej.patelczyk@intel.com>
Reviewed-by: Mika Kuoppala <mika.kuoppala@linux.intel.com>
Reviewed-by: Andi Shyti <andi.shyti@linux.intel.com>