INTEL_DII/FIXME: Restrict faulting to persistent vmas
Restrict faulting to persistent vmas unless fault is enabled across
using modparam enable_pagefault=1.

XXX:
The plan is to support faults on persistent vmas(vm_bind). There are
userspace tests which enable faults on a VM and still use legacy path
for submission. We need to formalize the api to return error to
userspace on such submissions.

Signed-off-by: Prathap Kumar Valsan <prathap.kumar.valsan@intel.com>
Reviewed-by: Niranjana Vishwanathapura <niranjana.vishwanathapura@intel.com>