INTEL_DII/FIXME: drm/i915/gt: Expose gt name in debugfs
With heterogeneous gt's (such as MTL), gt name is useful in identifying
which gt we are dealing with. Expose the name in debugfs.

FIXME: We wanted to add gt name to sysfs instead (see
http://intel-gfx-pw.fi.intel.com/series/12736/) but that is stalled because
at present no UMD has requested it. This may change in the future so adding
gt name to debugs may be temporary. On MTL having gt name in sysfs or
debugfs will help in identifying the SAMEDIA GT.

Reviewed-by: Umesh Nerlige Ramappa <umesh.nerlige.ramappa@intel.com>
Signed-off-by: Ashutosh Dixit <ashutosh.dixit@intel.com>