INTEL_DII: drm/i915/guc: Add FAST_REQUEST HXG ABI definition
In addition to the already defined REQUEST HXG message format,
which is used when sender expects some confirmation or data,
HXG protocol includes definition of the FAST REQUEST message,
that may be used when sender does not expect any useful data
to be returned.

Add ABI defines for this new message format, as GuC firmware
already has support for it.

v2: add missing #define

GuC: 64.0.0
Signed-off-by: Michal Wajdeczko <michal.wajdeczko@intel.com>
Acked-by: Piotr Piórkowski <piotr.piorkowski@intel.com>