drm/i915/gem: Skip vmap indirection for single block maps
When we enabled i915_gem_object_map() to use a vmap to handle
non-contiguous blocks of device memory, we left out the optimisation of
reusing the existing io_mapping for single contiguous blocks. However,
with the requirement for preallocation of page tables prior to execbuf
submission, we generate large volumes of small, single page objects that
are unused and immediately freed at the end of the submission path. This
introduces severe contention on the global vmap_area_lock and
free_vmap_area_lock. If we avoid using the vmap where we already have a
suitable io_mapping chunk, we can skip over the contention while
worrying about how to reduce the large number of useless preallocations.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Cc: Matthew Auld <matthew.auld@intel.com>
Cc: Thomas Hellström <thomas.hellstrom@intel.com>
Reviewed-by: Thomas Hellström <thomas.hellstrom@intel.com>