drm/i915/gem: Avoid allocations for looking up the first sg element
To allow random access to any scattergather element of the object, as
we iterate forward through the sg_table we build a radix cache. However,
that is overkill if we all we need, as we most frequently do, is just
the first sg element, which we can access immediately without any
additional locking, synchronisation or allocations.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Cc: Matthew Auld <matthew.auld@intel.com>
Cc: Thomas Hellström <thomas.hellstrom@intel.com>
Reviewed-by: Thomas Hellström <thomas.hellstrom@intel.com>