INTEL_DII: drm/i915/xehpsdv: Trim sg_table to match number of buddy chunks
Given that we cannot split the object more finely than
intel_memory_region.min_page_size, we know that there will be at most
that many chunks allocated in the buddy for the object. Instead of
estimating the maximum table size based on 4K pages, if we use a
worst-case estimate of 64K chunks we can reduce our preallocation by a
factor of 16.

Signed-off-by: Chris Wilson <chris.p.wilson@intel.com>
Tested-by: Oak Zeng <oak.zeng@intel.com>
Reviewed-by: Mika Kuoppala <mika.kuoppala@linux.intel.com>
Acked-by: Nirmoy Das <Nirmoy.das@linux.intel.com>