drm/i915/gem: Use XY_FAST_COLOR_BLT clears on gen12+
On DG1, XY_FAST_COLOR_BLT is substantially faster for larger operations
than the older XY_COLOR_BLT, and for our use case in clearing whole
page-aligned objects is functionally identical to the older command.
Let's take advantage of the faster instruction for when we zero out
pages on allocation.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Reviewed-by: Thomas Hellström <thomas.hellstrom@linux.intel.com>