drm/i915/guc: fix tasklet cleanup on shutdown
If a tasklet is scheduled, the tasklet_kill call waits for it to run
before killing it. This means that we need to make sure that the tasklet
can run, i.e. is enabled, even if the driver hit an error and wedged
beforehand.

v2: update guc_sched_engine_disabled as well (Matt)

v3: fix check in guc_sched_engine_disabled

Signed-off-by: Daniele Ceraolo Spurio <daniele.ceraolospurio@intel.com>
Cc: Matthew Brost <matthew.brost@intel.com>
Cc: John Harrison <John.C.Harrison@Intel.com>
Reviewed-by: Matthew Brost <matthew.brost@intel.com>