drm/i915: Wait for lmem initilisation completion
The firmware may not have fully completed initialising before the module
is loaded during early boot. We need to wait a short while to allow the
punit to complete its task. If we bail too early, we abort loading the
device leaving the system unusable. Simply wait until the device is
complete, or the user gives up. However, in the absence of having a
user, such as when builtin and loading during boot, we do need to impose
a timeout, where we have to err on the side of slow firmware.

The expectation is that it should take at most 50ms for the firmware to
initialise local memory, but given that the alternative is to abort
loading the device, allow waiting for a full 60s.

v2:(Tilak)
Expose and rename lmem init function

Fixes: c256af0d6c18 ("drm/i915/lmem: Fail driver init if LMEM training failed")
Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Cc: Matt Roper <matthew.d.roper@intel.com>
Cc: Matthew Auld <matthew.auld@intel.com>
Reviewed-by: Lucas De Marchi <lucas.demarchi@intel.com>
Signed-off-by: Tangudu Tilak <tilak.tangudu@intel.com>