INTEL_DII/NOT_UPSTREAM: vm_bind: Prevent eviction of active persistent BOs
During allocation in execbuf path, active BOs should not get evicted
under memory pressure as pages are reserved under the dma-resv.

But for persistent mappings, we are holding the BO lock only if the
module param enable_non_private_objects is set to true (which is not the
default value).
This module param will be removed once UMDs are OK with it and BO lock
will be held for all persistent BOs in the execbuff path.

Until then, bring back an old patch (below), which will skip evicting
any persistent BO if the VM's private reservation object is locked
(which will be in the execbuff path).
"INTEL_DII: drm/i915/vm_bind: Protect persistent vma unbind"

v2: Remove I915_GEM_OBJECT_UNBIND_TRYLOCK

Signed-off-by: Niranjana Vishwanathapura <niranjana.vishwanathapura@intel.com>
Acked-by: Prathap Kumar Valsan <prathap.kumar.valsan@intel.com>