drm/i915: Avoid assigning the last page of the GTT
Keep the last page unallocated, unless explicitly assigned by userspace,
prevent accidental prefetching past the end of the vm. If userspace is
using the last page by choice, we can assume that they have left a
sufficient gap in that page if being used as a batch buffer to avoid
stray execution. If instead we hand out the last page for an unknown
buffer, and it ends up being executed the prefetch beyond the end of the
vm may lead to a GPU hang.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Cc: stable@vger.kernel.org
Reviewed-by: Matt Roper <matthew.d.roper@intel.com>