INTEL_DII: drm/i915: define intel_partial_pages_for_sg_table
Function to retrieve the partial pages from the object, from mentioned
offset(pages). This is created as a subset of intel_partial pages to be
used for window blt copy feature which is introduced in forthcoming
patches.

This takes the sg_table to be filled in with pages and also passes out
the ptr to last scatterlist used. sg_table is trimmed based on the
parameter.

v2:
  PAGE_SHIFT is used along with blt related changes are removed. [Matt]
v3:
  trimming sg_table is left to the caller.
  gtt_view is dropped instead offset and size of view is passed in.
v4:
  Rebased on locking rework and obj->mm.lock removal. [Thomas]
v5:
  Rebased on upstream remap_contiguous_pages() refactoring. [Imre]

Signed-off-by: Ramalingam C <ramalingam.c@intel.com>
Acked-by: Matthew Auld <matthew.auld@intel.com>
Reviewed-by: CQ Tang <cq.tang@intel.com>