drm/i915/gem: Typecheck page lookups
We need to check that we avoid integer overflows when looking up a page,
and so fix all the instances where we have mistakenly used a plain
integer instead of a more suitable long. Be pedantic and add integer
typechecking to the lookup so that we can be sure that we are safe.

v2: Be even more pedantic and use pgoff_t as our page lookups must remain
compatible with the page cache, pgoff_t is currently exactly unsigned long.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Reviewed-by: Tvrtko Ursulin <tvrtko.ursulin@linux.intel.com>

(Rebased after mm lock removal patch series)