drm/i915/gt: Reduce frequency of checking LRC registers
Currently we check crucial ring registers inside the LRC, ones we don't
update normally as part of pinning, prior to submitting a request and
afterwards. This is very frequent, as we wanted to prevent coherency
issues when moving to discrete, but now that we have settled in, reduce
the debug checking frequency to the same rate as we check the context
image's redzone.

Signed-off-by: Chris Wilson <chris.p.wilson@intel.com>