drm/i915/gt: Only reset STOP_RING after a HW reset
Since we only poke the STOP_RING bit during a reset, we may reasonably
expect to only encouter that bit set then. It will not be part of the
default context image, and so not part of the initial context state for
each new context and so we can skip reseting the bit on every time we
pin the context for execution. We have a debug check, along with checking
the context redzone, to ensure that the bit is never set accidentally.

Signed-off-by: Chris Wilson <chris.p.wilson@intel.com>