INTEL_DII: drm/i915/mtl: implement wa 14015076503
The WA states that we need to alert the GSC FW before doing a GSC engine
reset and then wait for 200ms. The GuC will take care of this for engine
reset scenario, but i915 is still responsible for full GT reset, which
includes the GSC engine.

Given that we do full GT resets in the resume paths to cleanup the HW
state and that waiting 200ms in those scenarios would not be acceptable,
a faster path has been introduced where we try first to individually
reset the GuC and all engines except the GSC and only fall back to full
reset if that fails.

Note: the WA specs still mentions another possible shortcut were we
could only wait 15ms when resetting the GSC engine if the GSC FW is
idle. This however was not committed to by the GSC FW team yet.

Signed-off-by: Daniele Ceraolo Spurio <daniele.ceraolospurio@intel.com>
Cc: Matt Roper <matthew.d.roper@intel.com>
Cc: John Harrison <john.c.harrison@intel.com>
Reviewed-by: Rodrigo Vivi <rodrigo.vivi@intel.com>