INTEL_DII: drm/i915: Move address space ownership over to i915_drm_client
The client is a reference tracked struct, useful for when we may need to
check for the clients continued existence during a deferred vm shutdown,
where the drm_i915_file_private would no longer be available.

No one is currently accessing vm->file, and in the forthcoming patches,
the only use is with vm->file->client, risking the dangling pointer in
order to find resources stored on the client. We can cut out the
dangerous middle man and just store the client on the vm.

Signed-off-by: Chris Wilson <chris.p.wilson@intel.com>
Reviewed-by: Andi Shyti <andi.shyti@intel.com>