INTEL_DII/FIXME: drm/i915/guc: Dump error capture to dmesg on GuC init failure
There are some sporadic GuC init failures being observed by the CI system.
These are proving impossible to reproduce manually. The failure seems
to be pointing to an issue with receiving G2H responses - either a
missing interrupt or missing data when reading LMEM.

So add an error capture dump via dmesg on GuC init failure. This will
include both the GuC log and the contents of the CT channels.
Hopefully this will shed some light on what is going wrong.

Jira: VLK-23394

v2: Remove #if as this is now handled in the header. The message is also
    not needed as it's handled internally by intel_klog_error_capture()
    (Lucas)

Signed-off-by: John Harrison <John.C.Harrison@Intel.com>
Reviewed-by: Daniele Ceraolo Spurio <daniele.ceraolospurio@intel.com>