INTEL_DII: drm/i915/debugger: Use kfifo for event queue
With single event queued debugger is in lockstep with the
client. The aim is to be async wrt to client creating resources
with exception of events needing acks.

This lockstep with a queue of one has potential to hides
many concurrency issues and severely limits performance.

Use kfifo to allow more async behaviour. This also simplifies
the sending as the ownership of the event can be transferred
from the sender.

Signed-off-by: Mika Kuoppala <mika.kuoppala@linux.intel.com>
Acked-by: Maciej Patelczyk <maciej.patelczyk@intel.com>