drm/i915/gt: Avoid spinlock contention on timelines during retirement
When we have multiple threads flushing retired requests from the system
(for example when evicting from lmem we first flush retirement to free
up device memory), we find that those threads iterate through the
retirement list in lockstep. The spinlock for the iteration quickly
becomes the most contended lock (bouncing between cpus) on the system.
On larger systems with more clients and higher resource pressure, the
effect becomes more pronounced.

Though retiring the requests is not so critical, holding the spinlock
does prevent other threads on the system from submitting their requests
and making forward progress; it become a system-wide stall.

However, since we do not require the spinlock to iterate the list of
active timelines, we can skip it in favour of protecting the iteration
with rcu, removing a pleasing amount of code for the locking dance in
the process.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>