INTEL_DII: drm/i915/ATS: Handle faulting, when ATS lookups failed
ATS lookups may not yield an accessible page based on either of the
following scenarios:
a) The corresponding page is not resident in the CPU page tables
b) The page is resident, but the permissions are incorrect for the
attempted access (R/W mismatch)
In both cases, IOMMU would return a successful completion, and indicates
the permitted access types (R/W) - which for non-resident pages would be
R=0, W=0. In response, the device might either use Page Request Services
(PRS) to resolve the mapping directly through the IOMMU - which in turn
faults directly through the OS flows. However, if the device does not
support PRS, device can generate a fault into its own device driver to
resolved the faulting. Currently, Intel GPU's do not support PRS, so the
later flow is being used

JIRA: VLK-33096
Signed-off-by: Akeem G Abodunrin <akeem.g.abodunrin@intel.com>
Reviewed-by: Oak Zeng <Oak.Zeng@intel.com>