INTEL_DII: drm/i915/pf: Force VFs FLR while disabling VFs
GuC is unaware that VFs are being disabled and may continue to
wrongly assume that they are still in RUNNING state, which may
result in refusal of the unprovisioning or reprovisioning requests
from the PF unless VFs are properly FLR'ed.

Force FLR sequence for all VFs that were enabled, in case VMM
is not present or missed to do so.

Signed-off-by: Michal Wajdeczko <michal.wajdeczko@intel.com>