drm/i915: Move intel_guc_write_barrier() to a generic function
We have intel_guc_write_barrier() as a barrier to lmem writes. Move this
to a generic function as this is not specific to guc and we might need
to reuse this to ensure lmem writes across the barrier are completed
in-order.
As part of this change use an unused scratch pad register instead of a guc
specific register to get fencing. This register SOFTWARE_FLAGS_SPR33 is
going to be explicitly reserved for fencing. This makes the function safe
to call even before CT comuunication is enabled, because the register we
were using was also shared by i915 and GuC for the mmio based communication.
Also add a wmb() before the mmio write to ensure the WC buffers are flushed
so that the write to UC mapped mmio wont't pass them.

Signed-off-by: Prathap Kumar Valsan <prathap.kumar.valsan@intel.com>
Reviewed-by: Venkata Ramana Nayana <venkata.ramana.nayana@intel.com>
Acked-by: John Harrison <John.C.Harrison@Intel.com>