drm/i915: Add memory barrier before updating lrc tail pointer
For a non-multi lrc context, lrc tail pointer is updated by i915 to
submit the request. If the context is active and when gets scheduled in
the tail pointer gets restored from the lrca. We need to make sure the
writes to the ring is complete before tail pointer is updated in the
lrca, otherwise engine would end up executing invalid instructions from
the ring leading to a gpu hang.

Also now we are flushing ring and ct independently, remove the
assumption that LRC and CT vmas must be in the same region.

Signed-off-by: Prathap Kumar Valsan <prathap.kumar.valsan@intel.com>
Reviewed-by: Venkata Ramana Nayana <venkata.ramana.nayana@intel.com>
Acked-by: John Harrison <John.C.Harrison@Intel.com>